Week10_Cybersecurity/_my_finalproject/project/requirements.txt
google-api-python-client==2.83.0
google-auth==2.23.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
requests==2.31.0
